3a12c05: More size_t/int related compiler warning fixes
git-svn-id: https://forja.cenatic.es/svn/opendnie@410 52e26522-a1c9-48c4-a7af-28ea70901177

+++ opensc-opendnie/trunk/src/libopensc/cwa-dnie.c
@@ -211,7 +211,7 @@ int dnie_read_file(sc_card_t * card,
 	u8 *data;
 	char *msg = NULL;
 	int res = SC_SUCCESS;
-	unsigned int fsize = 0;	/* file size */
+	size_t fsize = 0;	/* file size */
 	sc_context_t *ctx = NULL;
 
 	if (!card || !card->ctx)
